+++ examples/resources/upcloud_server_group/import.sh
@@ -1,0 +1,1 @@
+terraform import upcloud_server_group.main ead4544f-10bf-42a3-b98a-a0fea2e2ad14