+++ internal/utils/utils.go
@@ -1,11 +1,14 @@
 package utils
 
 import (
+	"context"
+	"fmt"
 	"os"
 	"strings"
 	"time"
 
 	"github.com/UpCloudLtd/upcloud-go-api/v4/upcloud"
+	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
@@ -74,7 +77,7 @@ func WithRetry(fn func() (interface{}, error), retries int, delay time.Duration)
 	return nil, err
 }
 
-// ExpandStrings expands a terraform interface to slice of str
+// ExpandStrings expands a terraform list to slice of str
 func ExpandStrings(data interface{}) []string {
 	strSlice := []string{}
 	for _, s := range data.([]interface{}) {
@@ -82,6 +85,77 @@ func ExpandStrings(data interface{}) []string {
 	}
 
 	return strSlice
+}
+
+// SetOfStringsToSlice transforms a terraform set of strings to a slice of strings
+func SetOfStringsToSlice(ctx context.Context, data interface{}) ([]string, error) {
+	result := []string{}
+	providerErrMsg := "provider error: failed to transform set data"
+	debugLogPrefix := "transforming set of strings into slice failed;"
+
+	stringsSet, ok := data.(*schema.Set)
+	if !ok {
+		tflog.Debug(ctx, fmt.Sprintf("%s expected input data to be a schema.TypeSet but received %T", debugLogPrefix, data))
+		return result, fmt.Errorf(providerErrMsg)
+	}
+
+	for _, val := range stringsSet.List() {
+		valStr, ok := val.(string)
+		if !ok {
+			tflog.Debug(ctx, fmt.Sprintf("%s expected set elements to be of type string but received %T", debugLogPrefix, val))
+			return result, fmt.Errorf(providerErrMsg)
+		}
+
+		result = append(result, valStr)
+	}
+
+	return result, nil
+}
+
+// MapOfStringsToLabelSlice transforms a terraform map of strings to a LabelSlice
+func MapOfStringsToLabelSlice(ctx context.Context, data interface{}) (upcloud.LabelSlice, error) {
+	result := upcloud.LabelSlice{}
+	providerErrMsg := "provider error: failed to transform labels data"
+	debugLogPrefix := "transforming map of strings into labels slice failed;"
+
+	labelsMap, ok := data.(map[string]interface{})
+	if !ok {
+		tflog.Debug(ctx, fmt.Sprintf("%s expected input data to be a map of strings but received %T", debugLogPrefix, data))
+		return result, fmt.Errorf(providerErrMsg)
+	}
+
+	for k, v := range labelsMap {
+		value, ok := v.(string)
+		if !ok {
+			tflog.Debug(ctx, fmt.Sprintf("%s expected map elements to be of type string but received %T", debugLogPrefix, v))
+			return result, fmt.Errorf(providerErrMsg)
+		}
+
+		result = append(result, upcloud.Label{
+			Key:   k,
+			Value: value,
+		})
+	}
+
+	return result, nil
+}
+
+// LabelSliceToMap transorms `upcloud.LabelSlice` into a map of strings.
+// This can be used to set labels fetched from the API into a state
+func LabelSliceToMap(data upcloud.LabelSlice) map[string]string {
+	result := map[string]string{}
+	for _, label := range data {
+		result[label.Key] = label.Value
+	}
+
+	return result
+}
+
+// SliceOfStringToServerUUIDSlice converts slice of strings into `upcloud.ServerUUIDSlice`
+func SliceOfStringToServerUUIDSlice(strs []string) upcloud.ServerUUIDSlice {
+	result := make(upcloud.ServerUUIDSlice, len(strs))
+	copy(result, strs)
+	return result
 }
 
 // StorageAddressFormat takes the address in any format and extracts the bus